Improvement og global data storing
Changed the process of loading and saving global config data from/to the SD Card.
Now a temporary file is written before the other file is deleted.
If the loading of the right file fails, the other one is used and the original file is restored by the other one. This way there are no problems when power is lost during storing of the config data...